backend/requirements.txt

fastapi
uvicorn[standard]
pymongo[srv]==4.3.3
kafka-python==2.0.2
boto3==1.28.0
python-multipart==0.0.6
pydantic
python-dotenv
